feat(toc): add Toc component, hooks and story
Toc component exposes a Toc element and a Toc Provider that can be used to track ContentSections,
and to create an interactive navigation menu (toc) that uses intersectionobserver to keep track of
active position on page as well as scroll links to page content. The reason why we are not using
Headers is that Markdown does not provide a way for wrapping Title and content, so to enable
intersectionobserver to track content and not just headings we need to use ContentSections. With
Markdown the idea is to have one Markdown block per Content Section.

fix #7