feat(component api): refactor component api
Flex is now called Stack, and it has a deeper abstraction over FlexBox. I decided to offer
abstractions on some of the more complex/vague CSS concepts -- they can always be recreated with
Boxes or overrides if necessary. This api is not yet finished.

BREAKING CHANGE: Flex is now Stack.